fix: webauthn sign check

+++ tables/t_coupon_set_info.go
@@ -26,7 +26,7 @@ type CouponSetInfo struct {
 	Name          string          `gorm:"column:name;default:;NOT NULL"`
 	Note          string          `gorm:"column:note;default:;NOT NULL"`
 	Price         decimal.Decimal `gorm:"price:amount; type:decimal(50,10) NOT NULL DEFAULT '0' COMMENT '';"`
-	Num           int             `gorm:"column:num;default:0;NOT NULL"`
+	Num           int64           `gorm:"column:num;default:0;NOT NULL"`
 	BeginAt       int64           `gorm:"column:begin_at;default:0;NOT NULL"`
 	ExpiredAt     int64           `gorm:"column:expired_at;default:0;NOT NULL"`
 	Status        int             `gorm:"column:status;default:0;NOT NULL"`
